[contracts; remote] - v6 (#539)
* badoooms

* support upgrade from v6_x

* add check Key and MasterBlock

* add deploy func

* add checkshardblock

* add keyblock

* send token to  wrapper

* add object check

* update keyblock code

* add wrapper deploy

* update signature check

* [contracts] - update scripts

* change TvmCell to bytes

* add tvm.accept

* fix value in keyblock message to systemcontract.

* fix if

* fix index

* fix signature array

* fix signarute array

* fix return

* add same pubkey checks

* fix pubcheck

* fix

* fix tree value

* add destroy snapshots from other branch

* add check all commit parents

* add pin commit

* remove some snapshots for branch creating

* remove some snapshots for branch creating

* change snapshot salt

* fix pin commit

* update tree struct

* update iscorrect

* fix save

* fix deploy new snapshot

* update get functions

* update

* update time commit

* fix empty snap

* Copy 5.1 to 6

* update tree object

* update tree and snapshot

* update Tree object

* update hash calculate

* add value

* update tree address

* fix getter

* test

* test1

* test2

* test3

* maybe fix

* test4

* test5

* disable check sha in tree

* [contracts] - try to fix snapshot delete

* maybe fix

* fix

* update

* update

* update

* update hash

* fix

* turn off sendcontent

* destroy snaps after destroy branch

* update typer

* update big task

* update wallet to set approved small task

* fix for init upgrade

* first milestone

* fix answeris

* milestone2

* sort commits by time

* Revert order

* wip

* construct tree seems to work

* update snapshot for pin commit

* Fix getsnapshotaddr

* Push one commit works

* fix snap addr

* fix pintag

* fix answeris

* Push fetch for one commit works

* Removed some warnings

* rename snap getter

* fix approveSmallTask and change DeployIndex

* update Proposal list

* add getCellForIndex

* push works but not fetch

* Clone works for several commits

* Works deploy snapshot in branch

* try test on ci

* [web] - version 6.0.0

* Fix jenkins

* Fixed some tests

* fix diff addr in snapshot

* wip

* fix for Alex

* fix for Alex#2

* fix chain

* fix destroy snapshot

* change destroy snapshot system

* Fixed test

* [web] - fix blob delete

* add checktree after setcommit

* wip

* fix commit

* fix treeaccept message

* fix calc snap addr

* fix _number

* update isCorrect status

* [web] - update abis

* wip

* update type of tree check

* decrease size

* wip

* change proposals deploy

* fix remote

* Works with different file content

* fixed for files with the same content

* change proposal accept system

* old fix

* maybe fix

* [web] - v6 single proposal

* Merge branch 'contracts/new_version' of github.com:gosh-sh/gosh into contracts/new_version

* small fix

* fix Tree code in DiffC

* [web] - v6 diff check

* fix _isCorrect

* [web] - fix check diff

* update Commit deploy

* update react-gosh

* fix commit

* fix diff address

* fix

* add small changes

* small test

* final fix try 1

* update fix

* change diff

* perfect

* [web] - v6 fix push

* torn off check all parents

* turn on check all parents with initupgrade check

* fix one requirement

* add checkempty tree

* add correct message

* try to update

* try again

* try again

* try again

* try again

* Revert "try again"

This reverts commit 31d15790f0ca3ba7ecbf7ba22d8192d7966c731e.

* Revert "try again"

This reverts commit ebbbe2156378e709d98629b75b72733d236c52be.

* Revert "try again"

This reverts commit 9d05e28423eaa04496dddc0298f38db9004e8fde.

* Revert "try again"

This reverts commit 8a9b52348565fa35e5dafc85c3bae93ffebfca5c.

* Revert "try to update"

This reverts commit 672d148ef28a977c7ab10f9c99694bdd4f500191.

* update modifiers

* [web] - update prod envs with v6

* Merge remote v6

---------

+++ tests/27_snap_check_readiness_before_create_branch.test.sh
@@ -13,6 +13,11 @@ if [[ "$VERSION" == *"v1_x"* ]]; then
   exit 0
 fi
 
+if [[ "$VERSION" == *"v6_x"* ]]; then
+  echo "Test is ignored for v6 because in v6 we don't deploy snapshots for a new branch"
+  exit 0
+fi
+
 REPO_NAME="repo27_$(date +%s)"
 
 [ -d $REPO_NAME ] && rm -rf $REPO_NAME
